Update NPM Version Lookup (#175)
The previous version of the script used `npm view` to check for the 
existence of a package version.

If the version did not exist, this resulted in an error instead of
no output as expected by the remainder of the script.

Instead, use `npm show` and grep the result. This correctly outputs 
a line of text for a version that exists and nothing for a version that 
does not.