SW-2625 Build a Multiple-Selection Dropdown Component
A multiple selection dropdown component is build here. It expects a map
of identifiers (generic) to values (strings) for its options. It also
expects a onAdd and onRemove handlers for the consumer to update its
selectedOptions with.

The component shows pills for selected options with a remove button, and
the remaining options are shown in a dropdown. The onAdd handler is
called with the map key whenever one of these is clicked.