SW-2625 Build a Multiple-Selection Dropdown Component (#270)
A multiple selection dropdown component is build here. It expects a map
of identifiers to values for its options. It expects a onAdd and onRemove
handlers for the consumer to update its selectedOptions with. A value-
Renderer function is required to convert the values to a renderable string.

The component shows pills for selected options with a remove button, and
the remaining options are shown in a dropdown. The onAdd handler is
called with the map key whenever one of these is clicked.