Unfortunately, due to faulty logic, the last attempt at this required a double click to close the Popover if the MultiSelect had been focused and blurred within the Popover. This refactors it to make it make a little more sense (hopefully), by implementing an onFocus prop to the MultiSelect and capturing the click event that propagates from clicking away from the Popover.
The implementer of this MultiSelect-within-Popover-with-desired-blur/closing-behavior unfortunately needs to implement a cheeky setTimeout for the functionality to work as expected. This is because the state of the component updates from the onBlur event resolution before the onClick event resolves, resulting in an impossible (at least through my many attempts) situation to achieve the desired effect.