SW-7130 Map legend component (#676)
This unifies a some map legend funcitonalities that have been scattered on terraware-web.

Most of the work is separating props into 3 distinct types: layers, features and highlights. Layers and highlights are geometric shapes drawn on the map. Features are pins with icons placed onto the map. This should remove some cases where invalid props are used for legends. This should also serve as a good framework for us to add more features (for example showing different kinds of features, multiple layers, etc.)

[Screen Recording 2025-07-22 at 5.05.41 PM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/NOxkRPuhAfT4F7nmdXtG/c9a51cbe-50e0-4b03-9f46-95e3179ff2f2.mov" />](https://app.graphite.dev/media/video/NOxkRPuhAfT4F7nmdXtG/c9a51cbe-50e0-4b03-9f46-95e3179ff2f2.mov)

(The recording software messes with the pointer cursor. But this work correctly on storybook).