Implemented markers on Map (#681)
[Screen Recording 2025-08-04 at 12.38.02 PM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/NOxkRPuhAfT4F7nmdXtG/8d26f8a5-3a3d-4876-8397-1c4d25bb013c.mov" />](https://app.graphite.dev/media/video/NOxkRPuhAfT4F7nmdXtG/8d26f8a5-3a3d-4876-8397-1c4d25bb013c.mov)

This will be used to display photos and plants on the Map. 

Clustering and numbering to be implemented.

The demo video showed a nifty tool to place markers and also handling on click events. The onClick event destroys the marker but other use cases would include opening MapDrawer to display a photo, or zoom-into the cluster.