Added map marker clustering functionality (#682)
[Screen Recording 2025-08-04 at 7.48.11 PM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/NOxkRPuhAfT4F7nmdXtG/2714aa23-f145-4246-955d-88fc720450a0.mov" />](https://app.graphite.dev/media/video/NOxkRPuhAfT4F7nmdXtG/2714aa23-f145-4246-955d-88fc720450a0.mov)

Implemented clustering via pixel distance. This is essentially a lightweight implementation of Mapbox supercluster. It can likely handle up to a few hundred points but will not be as scalable.

This is implemented this way to preserve functions of our React actions and hooks, as well as our component library. Using Mapbox built-in cluster functions will require a new set of styles specific to Mapbox canvases.